Initial commit of the package
This is the initial commit of the package; this includes all of the core
commands, sample key and mouse bindings, a quick readme, the actual
plugin code itself and resources to add the usual commands to the main
menu and the command palette.

We currently implement an event listener that will rewrite the normal
font commands into our commands, so that even the default commands will
respect the new reset method and font size contraints.