Include a command to set explicit font size
This commit extends the internal command set so that along with a simple
font resize and reset, the font size can also be specifically set to a
given size directly.

This works for all of the modes that we can change the font size for,
and is predicated on the idea that you might want to have a way to jump
to a specific size (e.g. when doing a screen share or similar).